Fix server startup on NetBSD (and probably other systems)
my_open_parent_dir_nosymlinks() calls openat() with dfd = -1, which
Linux takes for root of filesystem, but this behavior is not standard
as per OpenGroup XSH:
http://pubs.opengroup.org/onlinepubs/9699919799/functions/open.html

On NetBSD, dfd = -1 causes openat() to fail with EBADF, which breaks
server startup. The same result probaly happens on any non Linux
system implementing openat().

The proposed fix is to call plain open() when dfd = -1