MDEV-16986 Unitialized mutex, SIGSEGV and assorted assertion failures in Aria code
The problem was that when a mysql.proc table was
opened for reading it was added to the current Aria
transaction context but never properly deleted from
it. Normally this isn't a problem, except if the
mysql.proc table is closed before the transaction
ended, which happened in this test case.

Fixed by removing mysql.proc from the transaction
context before closing the table.