MDEV-16250: Spider system tables are MyISAM and needs repair after a crash
During a server restart that follows a server crash, there may be error
messages that indicate that certain system tables are marked as crashed and
may be corrupted.  Upon checking the system tables that are marked as crashed,
it may be found that there is no corruption that needs repair.  However, the
error messages that are issued imply that the user needs to nevertheless
manually repair the system tables that were marked as crashed.  These issues
have been moved to a separate bug MDEV-17068.

MDEV-16250 addresses the work to make the Spider system tables crash safe.
This work involves the following changes in 10.4:
- Changes to the install_spider.sql script to change the storage engine for
  the Spider system tables to Aria.  This is implemented in such a way as to
  allow the script to be run repeatedly without any harm or errors.
- Changes to the init_spider.inc script that is run during initialization of
  every test in the Spider test suites.  This script now uses the Aria
  storage engine for the Spider system tables in MariaDB Server 10.4 and
  later releases.
- Added a test to the Spider test suite to display the storage engine of each
  Spider system table, to verify that the correct storage engine is used.

Author:
  Jacob Mathew.

Reviewer:
  Kentoku Shiba.