Revert "Added -j option to dpkg-buildpackage to speed up build"
This reverts commit a89a6e4f774d9b01bee0d33c0b2f0d6a1dfae39d.

We cannot blindly assume all CPUs on all hosts are available for
building. For example our own buildbot.mariadb.org builders go
amok if everey host CPU core is used in parallel for every build
on a host.

This option should be configurable, and in fact it already is via
the parallel=N setting in DEB_BUILD_OPTIONS. See debian/rules
for details.