MDEV-17016: failed assertion in the ha_partition::external_lock()
This patch fixes some partition-related regressions and
failures under rare conditions. In particular, unreleased
auto-increment lock leds to assertion failure when mysqld
compiled in DEBUG mode (only in it) and when test uses
partitions and generates keys outside the allowed range.

A new test has been added, which reliable replays and
tests this failure. Two other tests are improved to
increase the reproducibility of the results and to
avoid false failures when the second node starts earlier
than the first one.

In particular, several partitions-related mtr tests fail
because the write_record() function (which located in the
sql_insert.cc file) goes to the "before_trg_err" label,
but after failure in the update_auto_increment() method
the insert_id_for_cur_row field will become zero. As a
result, the subsequent call of the restore_auto_increment()
function does not restore value of the next_insert_id
field, which in the future results in an error in the
release_auto_increment method.

To prevent this, we need to modify release_auto_increment
method to release multi-row statement lock taken in
the get_auto_increment even if the field next_insert_id
contains zero.

Also, in sql_insert.cc (in the write_record function),
we need to explicitly check that the generated auto-increment
value is consumed by UPDATE and not in the range reserved
for another row.

+cosmetic change in the ha_innodb.cc (one of the assertions
has already been moved to another place and can be deleted)

https://jira.mariadb.org/browse/MDEV-17016