Use accessors rec_is_metadata(), rec_is_alter_metadata()
rec_info_bits_valid(): Remove. The only case when this would fail
is when the record is the generic ALTER TABLE metadata record.

rec_is_alter_metadata(): Check if a record is the metadata record
for instant ALTER TABLE (other than ADD COLUMN). NOTE: This function
must not be invoked on node pointer records, because the delete-mark
flag in those records may be set (it is garbage), and then a debug
assertion could fail because index->is_instant() does not necessarily
hold.

rec_offs_validate(): FIXME: Temporarily disabled an assertion
that does not hold if rec_is_alter_metadata(). Because we do not know
here if a ROW_FORMAT=REDUNDANT record is for the leaf page, we cannot
invoke rec_is_alter_metadata() here.