Revert changes to rec_offs_data_size() and friends
Do not exclude instantly dropped columns from the size of the record.
This has to be handled at the upper level, by row_build() and the
to-be-written replacement of row_build_index_entry(), called
row_build_clust_index_entry().

Hopefully only a small number of functions will have to directly
access those clustered index record fields that can be reordered
or instantly dropped.