MDEV-17466: Remove the debug assertion
This reverts commit 2d4075e1d9dba3dd8503d65fdb145a6dcc056c51
where the debug assertion was added. There seems to be a potential
problem in the purge of indexes that depend on virtual columns.

Ultimately, we should change the InnoDB undo log format so that
all actual secondary index keys are stored there, also for
virtual or spatial indexes. In that way, purge and rollback would
be more straightforward.