Fixing MW-286 test
The MW-286 mtr test was not deterministic. The idead in the test was that ALTER execution
in node 2 would be BF aborted by INSERT replication from node 1.
However, with mariaDB, the ALTER is pending in MDL lock wait loop, and avoids the MDL lock
conflict with replicated INSERT.
In this patch, the mtr test is refactored, by starting ALTER in node 2 first. The ALTER execution
is stopped in debug sync point after exclusive MD lock was granted for ALTER.
After that a single INSERT statement replication from node, will cause BF abort for ALTER.

When testing with the refactored MW-286 test, it turned out that MariaDB will trigger an assert
when BF aborting the ALTER execution. The assert is because of double locking lock_sys mutex.
This issue has been fixed in ha_innodb.cc, to mark the victim transaction early with
'trx::was_chosen_as_deadlock_victim' flag. When victim has this flag set, we know that BF abort is
ongoing and that lock_sys mutex is already grabbed.