Initial step at fixing read of old NOT NULL columns
We want dict_index_t::n_nullable to reflect the original NOT NULL
state of the columns.

dict_table_t::init_instant(): Remove an assertion that would fail
when the InnoDB data dictionary is out of sync with the MariaDB
data dictionary. Correctly compute n_nullable. (With this,
the assertion ut_ad(index.n_nullable == n_nullable) at the
end of the function will fail.)